Remove unused and unreferenced "canonicalization views"
These are now handled in the canonicalization model and admin
