+++ pyproject.toml
@@ -75,8 +75,8 @@ version-file = "src/nomnom/_version.py"
 [tool.hatch.build.targets.wheel]
 packages = ["src/nomnom"]
 
-[tool.uv]
-dev-dependencies = [
+[dependency-groups]
+dev = [
     "pytest-django~=4.7",
     "pytest~=8.0",
     "django-stubs~=4.2",
@@ -111,6 +111,7 @@ dev-dependencies = [
     "pytest-md-report>=0.6.3",
 ]
 
+[tool.uv]
 override-dependencies = [
     "django-celery-beat~=2.5",
     "django~= 5.1",
@@ -153,7 +154,7 @@ exclude_also = [
 profile="django"
 
 [tool.django-stubs]
-django_settings_module = "config.settings"
+django_settings_module = "nomnom.test_settings"
 
 [tool.ruff]
 exclude = [
